feat: 岗位管理和岗位级别管理页面完成

+++ init.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS `levels`
 (
     `id`      int     NOT NULL AUTO_INCREMENT,
     `level`   char(8) NOT NULL,
-    `payment` int     NOT NULL,
+    `payment` double  NOT NULL,
     PRIMARY KEY (`id`),
     UNIQUE INDEX `level` (`level`) USING BTREE
 );
